+++ tailscale/rootfs/etc/s6-overlay/scripts/stage2_hook.sh
@@ -65,6 +65,7 @@ fi
 # Remove unused options
 healthcheck_offline_timeout=$(bashio::jq "${options}" '.healthcheck_offline_timeout | select(.!=null)')
 healthcheck_restart_timeout=$(bashio::jq "${options}" '.healthcheck_restart_timeout | select(.!=null)')
+forward_to_host=$(bashio::jq "${options}" '.forward_to_host | select(.!=null)')
 if bashio::var.has_value "${healthcheck_offline_timeout}"; then
     bashio::log.info 'Removing deprecated healthcheck_offline_timeout option'
     bashio::addon.option 'healthcheck_offline_timeout'
@@ -72,6 +73,10 @@ fi
 if bashio::var.has_value "${healthcheck_restart_timeout}"; then
     bashio::log.info 'Removing deprecated healthcheck_restart_timeout option'
     bashio::addon.option 'healthcheck_restart_timeout'
+fi
+if bashio::var.has_value "${forward_to_host}"; then
+    bashio::log.info 'Removing deprecated forward_to_host option'
+    bashio::addon.option 'forward_to_host'
 fi
 
 # Update changed options
@@ -106,10 +111,8 @@ then
     rm /etc/s6-overlay/s6-rc.d/post-tailscaled/dependencies.d/local-network
 fi
 
-# Disable forwarding service when userspace-networking is enabled or forwarding to host is disabled
-if bashio::config.true "userspace_networking" || \
-    bashio::config.false "forward_to_host"
-then
+# Disable forwarding service when userspace-networking is enabled
+if bashio::config.true "userspace_networking"; then
     rm /etc/s6-overlay/s6-rc.d/user/contents.d/forwarding
 fi